Create README.md
This project was developed during an internship at Afame Technologies, where I worked as a Machine Learning Intern. The goal of this project is to create a model that can accurately detect spam emails using a Naive Bayes classifier. The model achieves an impressive 98% accuracy on the spam detection dataset.